Add LLMCallsRecordable concern
This stores the functionality which assigns LLM responses and metrics
into a concern, allowing it to be reused across models.

We know that Answer and AnswerAnalysis both need this functionality and
likely more modeles will in the future so it makes sense to abstract it.
