+++ app/models/answer_analysis.rb
@@ -1,3 +1,5 @@
 class AnswerAnalysis < ApplicationRecord
+  include LlmCallsRecordable
+
   belongs_to :answer
 end
