+++ spec/models/answer_analysis_spec.rb
@@ -1,0 +1,5 @@
+RSpec.describe AnswerAnalysis do
+  include_examples "llm calls recordable" do
+    let(:model) { build(:answer_analysis) }
+  end
+end
